+++ Source/UEProjectIntelligence/Private/Operations/UEPIAnimationOperations.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Edit/UEPIEditOperation.h"
+
+namespace UE::ProjectIntelligence
+{
+	TSharedRef<IUEPIEditOperation> MakeUEPIAnimationOperation(const FUEPIEditOperationDescriptor& Descriptor);
+}